fix next-auth type errors
For some reason it breaks for me when the next-auth.d.ts file was named
the same as next-auth import.
When i hovered over the `from 'next-auth'` it showed
a different path that was outside node_modules,
whereas all other imports (which were working)
were showing coming from node_modules.

I'm not sure why it's behaving like this; the tsconfig.js
here looks basically the same as the next-auth example. I don't
understand atm, but anyway, if the file is named differently then all seems fine,
and the JWT module extension seems to work fine still too,
so I'm happy enough for now.