dont import from root, prevent import clashing filenames
This is actually what was happening in my previous
commit; it was trying to import next-auth from root project
first, and it found a file that matched that, which takes priority over
matches found in node_modules. Don't do that anymore, it now seems
annoying and like a footgun. Import things relatively too unfortunately,
which is annoying, but at least it's less footgunny.